feat(api-file-manager-server): implement upload abstractions, remove old schema
Replace the legacy GraphQL upload resolvers with three DI feature
implementations: GetUploadPayload, CreateMultiPartUpload, and
CompleteMultiPartUpload. Use Request abstraction from @webiny/handler
to access the Fastify request for resolveServerUrl. Remove the old
graphql/schema.ts, multiPartUpload/ classes, and utils/getPresignedPostPayload.ts.
Drop unused @webiny/handler-graphql and p-map dependencies.
