+++ packages/api-file-manager-server/src/features/CompleteMultiPartUpload/feature.ts
@@ -1,0 +1,9 @@
+import { createFeature } from "@webiny/feature/api";
+import { CompleteMultiPartUploadUseCaseImplementation } from "./CompleteMultiPartUploadUseCase.js";
+
+export const CompleteMultiPartUploadFeature = createFeature({
+    name: "FileManagerServer/CompleteMultiPartUpload",
+    register(container) {
+        container.register(CompleteMultiPartUploadUseCaseImplementation);
+    }
+});
