+++ packages/api-file-manager-server/src/features/CreateMultiPartUpload/feature.ts
@@ -1,0 +1,9 @@
+import { createFeature } from "@webiny/feature/api";
+import { CreateMultiPartUploadUseCaseImplementation } from "./CreateMultiPartUploadUseCase.js";
+
+export const CreateMultiPartUploadFeature = createFeature({
+    name: "FileManagerServer/CreateMultiPartUpload",
+    register(container) {
+        container.register(CreateMultiPartUploadUseCaseImplementation);
+    }
+});
