+++ packages/api-file-manager-server/src/features/GetUploadPayload/feature.ts
@@ -1,0 +1,9 @@
+import { createFeature } from "@webiny/feature/api";
+import { GetUploadPayloadUseCaseImplementation } from "./GetUploadPayloadUseCase.js";
+
+export const GetUploadPayloadFeature = createFeature({
+    name: "FileManagerServer/GetUploadPayload",
+    register(container) {
+        container.register(GetUploadPayloadUseCaseImplementation);
+    }
+});
