+++ packages/api-file-manager-server/src/index.ts
@@ -1,7 +1,6 @@
 import { existsSync } from "node:fs";
 import { mkdirSync } from "node:fs";
 import { ContextPlugin } from "@webiny/api";
-import { createServerGraphQLSchema } from "./graphql/schema.js";
 import { uploadRoutesPlugin } from "~/routes/uploadRoutes.js";
 import { modifyFastifyPlugin } from "~/routes/uploadRoutes.js";
 import { CleanupStaleMultipartUploadsFeature } from "~/features/CleanupStaleMultipartUploads/feature.js";
@@ -10,6 +9,9 @@ import { ExtractMetadataFeature } from "~/features/ExtractMetadata/feature.js";
 import { FlushCacheFeature } from "~/features/FlushCache/feature.js";
 import { GetFileContentsByIdFeature } from "~/features/GetFileContentsById/feature.js";
 import { GetFileContentsByKeyFeature } from "~/features/GetFileContentsByKey/feature.js";
+import { GetUploadPayloadFeature } from "~/features/GetUploadPayload/feature.js";
+import { CreateMultiPartUploadFeature } from "~/features/CreateMultiPartUpload/feature.js";
+import { CompleteMultiPartUploadFeature } from "~/features/CompleteMultiPartUpload/feature.js";
 export { createFileUploadModifier } from "@webiny/api-file-manager/features/upload/index.js";
 export { createAssetDelivery } from "./assetDelivery/createAssetDelivery.js";
 
@@ -39,6 +41,9 @@ const contextPlugin = new ContextPlugin(context => {
     ExtractMetadataFeature.register(container);
     GetFileContentsByIdFeature.register(container);
     GetFileContentsByKeyFeature.register(container);
+    GetUploadPayloadFeature.register(container);
+    CreateMultiPartUploadFeature.register(container);
+    CompleteMultiPartUploadFeature.register(container);
     CleanupStaleMultipartUploadsFeature.register(container);
 });
 
@@ -46,7 +51,6 @@ contextPlugin.name = `fileManagerServer.context`;
 
 export const createFileManagerServer = () => [
     contextPlugin,
-    createServerGraphQLSchema(),
     uploadRoutesPlugin,
     modifyFastifyPlugin
 ];
